refactor: View 패키지에 맞게 ViewModel 패키지 구조 변경
