+++ presentation/src/main/res/layout/item_my_recipe.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="myRecipeViewModel"
-            type="com.kdjj.presentation.viewmodel.my.MyRecipeViewModel" />
+            type="com.kdjj.presentation.viewmodel.home.my.MyRecipeViewModel" />
 
         <variable
             name="myRecipeItem"
