+++ presentation/src/main/res/layout/item_my_recipe_add_recipe.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="myRecipeViewModel"
-            type="com.kdjj.presentation.viewmodel.my.MyRecipeViewModel" />
+            type="com.kdjj.presentation.viewmodel.home.my.MyRecipeViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
